Multi-pointer - the sum of four numbers
Given an array nums consisting of n integers and a target value target. Please find and return the tuples [nums[a], nums[b], nums[c], nums[d]] that meet all of the following conditions and are not duplicated (if the elements of two tuples correspond one-to-one, then the two tuples are considered duplicated) :



0 <= a, b, c, d < n

a, b, c and d are not the same

nums[a] + nums[b] + nums[c] + nums[d] == target

You can return the answers in any order.