Multi-pointer - Verify palindrome strings
f all uppercase characters are converted to lowercase and all non-alphanumeric characters are removed, the phrase reads the same whether it is read forward or backward. Then it can be regarded as a palindrome string.



Both letters and numbers belong to alphanumeric characters.



Given a string s, if it is a palindrome, return true; Otherwise, return false.