Multi-pointer plant watering
Alice and Bob are going to water the n plants in the garden. Plants are lined up in a row and marked from left to right, with numbers ranging from 0 to n - 1. Among them, the position of the i-th plant is x = i.



Each plant needs a specific amount of water. Alice and Bob each have a water jug, which was full at first. They completed watering in the way described below:



Alice waters the plants from left to right, starting from plant 0. Bob watered the plants in the order from right to left, starting from plant n - 1. They water the plants simultaneously.

No matter how much water is needed, the time required to water each plant is the same.

If the water in Alice/Bob's water tank is sufficient to fully irrigate the plants, they must water the plants. Otherwise, they first (immediately) refill the jars and then water the plants.

If Alice and Bob reach the same plant, the person with more water in the current water tank will water this plant. If they have the same amount of water, then Alice will water this plant.

Given an integer array plants with indices starting from 0, the array consists of n integers. Among them, plants[i] represents the amount of water required by the i-th plant. There are also two integers, capacityA and capacityB, which respectively represent the capacities of Alice and Bob's water tanks. Return the number of times the two people refilled the water tank during the process of watering all the plants.