Sliding window - Longest half-repeated string
Given a string s with indices starting from 0, this string only contains numeric characters from 0 to 9.



If at most one pair of adjacent characters in a string t are equal, then this string t is called semi-repetitive. For example, "0010", "002020", "0123", "2002" and "54944" are semi-repeated strings However, "00101022" (where adjacent pairs of the same number are 00 and 22) and "1101234883" (where adjacent pairs of the same number are 11 and 88) are not semi-repeated strings.



Please return the length of the longest half-repeated substring in s.