Sliding window - The minimum operand for subtracting x to 0
Given an integer array nums and an integer x. Each time you perform an operation, you should remove the leftmost or rightmost element of the array nums and then subtract the value of that element from x. Please note that the array needs to be modified for the subsequent operations.



If x can be exactly reduced to 0, return the minimum operand; Otherwise, return -1.