Dichotomy - the number of successful spells and potions
Given two positive integer arrays, spells and potions, with lengths of n and m respectively, spells[i] represents the energy intensity of the i-th spell, and potions[j] represents the energy intensity of the j-th potion.



At the same time, give you an integer "success". If the product of the energy intensity of a spell and a potion is greater than or equal to success, then they are regarded as a successful combination.



Please return an integer array pairs of length n, where pairs[i] is the number of potions that can be successfully combined with the i-th spell.