Binary h-index Ⅱ
Given an integer array citations, where citations[i] represents the number of citations of the researcher's i-th paper, citations have been arranged in non-descending order. Calculate and return the H-index of this researcher.



The definition of the H-index: h stands for "high citations". A researcher's H-index refers to the fact that at least h of his or her (n papers) have been cited at least h times.



Please design and implement an algorithm with logarithmic time complexity to solve this problem.

以上翻译结果来自有道神经网络翻译（YNMT）· 通用场景