Binary spiral array minimum value
Given an array of length n, which is pre-arranged in ascending order and rotated 1 to n times, the input array is obtained. For example, the original array nums = [0,1,2,4,5,6,7] might result in:

If rotated four times, then [4,5,6,7,0,1,2] can be obtained.

If rotated 7 times, then [0,1,2,4,5,6,7] can be obtained.

Note, array [a [0], a [1], a [2],... and a [n - 1]] array to the result of the rotation a [a] [n - 1, a [0], a [1], a [2],..., a [n - 2]].



Given an array 'nums' with distinct element values, it was originally an ascending array and has undergone multiple rotations as described above. Please find and return the smallest element in the array.



You must design an algorithm with a time complexity of O(log n) to solve this problem