Linked list _ Repeat k groups of reversed linked lists
Given the head node head of your linked list, flip every k nodes in groups. Please return to the modified linked list.



k is a positive integer whose value is less than or equal to the length of the linked list. If the total number of nodes is not an integer multiple of k, then please keep the remaining nodes in their original order.



You cannot merely change the values inside the nodes; instead, you need to actually perform node swaps.