Linked list - Judge the linked list
Given a head node head of a linked list, determine whether there is a cycle in the list.



If there is a node in the linked list that can be reached again by continuously tracking the next pointer, then there is a loop in the linked list. To represent the cycle in a given linked list, the evaluation system internally uses the integer pos to indicate the position where the tail of the linked list is connected to the list (the index starts from 0). Note: pos is not passed as a parameter. It is merely to identify the actual situation of the linked list.



If there are cycles in the linked list, return true. Otherwise, return false.