Linked list - Determine the link list and the starting point
Given the head node head of a linked list, return the first node from which the list begins to enter the loop. If the linked list is acyclic, null is returned.



If there is a node in the linked list that can be reached again by continuously tracking the next pointer, then there is a loop in the linked list. To represent the cycle in a given linked list, the evaluation system internally uses the integer pos to indicate the position where the tail of the linked list is connected to the list (the index starts from 0). If pos is -1, there is no cycle in this linked list. Note: pos is not passed as a parameter; it is merely used to identify the actual situation of the linked list.



The linked list is not allowed to be modified.