Linked list - maximum twin sum
In a linked list of size n and where n is an even number, for i where 0 <= i <= (n / 2) -1, the twin node of the i-th node (with indices starting from 0) is the (n-1-i) th node.



For instance, if n = 4, then node 0 is a twin of node 3, and node 1 is a twin of node 2. These are all the twin nodes in the linked list of length n = 4.

Twin sum is defined as the sum of the values of a node and its twin nodes.



Given the head node head of a linked list with an even length, please return the maximum twin sum of the linked list.